+++ src/us/kbase/workspace/kbase/KBaseIdentifierFactory.java
@@ -200,10 +200,12 @@ public class KBaseIdentifierFactory {
 		for (final ObjectSpecification o: objects) {
 			if (o == null) {
 				throw new NullPointerException(
-						"Objects in the object specification list cannot be null");
+						"Objects in the object specification list cannot "+
+						"be null");
 			}
 			final ObjectIdentifier oi;
 			try {
+				mutateObjSpecByRefString(o);
 				oi = processObjectIdentifier(
 					new ObjectIdentity()
 						.withWorkspace(o.getWorkspace())
@@ -218,83 +220,165 @@ public class KBaseIdentifierFactory {
 						"Error on ObjectSpecification #"
 						+ objcount + ": " + e.getLocalizedMessage(), e);
 			}
-			final ObjectPaths paths;
-			if (o.getIncluded() != null && !o.getIncluded().isEmpty()) {
-				paths = new ObjectPaths(o.getIncluded(),
-						longToBoolean(o.getStrictMaps(),
-								ObjectPaths.STRICT_MAPS_DEFAULT),
-						longToBoolean(o.getStrictArrays(),
-								ObjectPaths.STRICT_ARRAYS_DEFAULT));
-			} else {
-				paths = null;
-			}
-			final List<ObjectIdentifier> refchain;
-			try {
-				refchain = processRefChains(o);
-			} catch (IllegalArgumentException e) {
-				throw new IllegalArgumentException(
-						"Error on ObjectSpecification #"
-						+ objcount + ": " + e.getLocalizedMessage(), e);
-			}
-			if (paths == null && refchain == null) {
-				objs.add(oi);
-			} else if (paths == null) {
-				objs.add(new ObjectIDWithRefChain(oi, refchain));
-			} else {
-				objs.add(new ObjIDWithChainAndSubset(oi, refchain, paths));
-			}
+			objs.add(buildObjectIdentifier(o, oi, objcount));
 			objcount++;
 		}
 		return objs;
 	}
 
-	private static List<ObjectIdentifier> processRefChains(
+	private static ObjectPaths getObjectPaths(final ObjectSpecification o) {
+		if (o.getIncluded() != null && !o.getIncluded().isEmpty()) {
+			return new ObjectPaths(o.getIncluded(),
+					longToBoolean(o.getStrictMaps(),
+							ObjectPaths.STRICT_MAPS_DEFAULT),
+					longToBoolean(o.getStrictArrays(),
+							ObjectPaths.STRICT_ARRAYS_DEFAULT));
+		}
+		return null;
+	}
+	
+	private static ObjectIdentifier buildObjectIdentifier(
+			final ObjectSpecification o,
+			ObjectIdentifier oi,
+			final int objcount) {
+		final ObjectPaths paths = getObjectPaths(o);
+		final List<ObjectIdentifier> refchain;
+		try {
+			final RefChainResult res = processRefChains(o);
+			if (res == null) {
+				refchain = null;
+			} else if (res.isToPath) {
+				refchain = new LinkedList<>();
+				if (res.path.size() > 1) {
+					refchain.addAll(res.path.subList(1, res.path.size() - 1));
+				}
+				refchain.add(oi);
+				oi = res.path.get(0);
+			} else {
+				refchain = res.path;
+			}
+		} catch (IllegalArgumentException e) {
+			throw new IllegalArgumentException(
+					"Error on ObjectSpecification #"
+					+ objcount + ": " + e.getLocalizedMessage(), e);
+		}
+		if (paths == null && refchain == null) {
+			return oi;
+		} else if (paths == null) {
+			return new ObjectIDWithRefChain(oi, refchain);
+		} else {
+			return new ObjIDWithChainAndSubset(oi, refchain, paths);
+		}
+	}
+
+	private static void mutateObjSpecByRefString(
 			final ObjectSpecification o) {
-		final List<ObjectIdentity> objrefs;
+		if (o.getRef() == null || o.getRef().trim().isEmpty()) {
+			return;
+		}
+		final String[] reflist = o.getRef().trim().split(";");
+		o.setRef(reflist[0].trim());
+		final List<String> refs = new LinkedList<>();
+		for (int i = 1; i < reflist.length; i++) {
+			refs.add(reflist[i].trim());
+		}
+		if (!refs.isEmpty()) {
+			if (countRefPathsInObjectSpec(o) > 0) {
+				throw new IllegalArgumentException(
+						"Only one of the 5 options for specifying an "+
+						"object reference path is allowed");
+			}
+			o.setObjRefPath(refs);
+		}
+	}
+		
+	private static int countRefPathsInObjectSpec(final ObjectSpecification o) {
+		int count = 0;
+		// there's probably some obnoxiously smarter way to do this
 		if (o.getObjPath() != null && !o.getObjPath().isEmpty()) {
-			objrefs = o.getObjPath();
-		} else {
-			objrefs = null;
+			count++;
 		}
-		final List<String> refs;
 		if (o.getObjRefPath() != null && !o.getObjRefPath().isEmpty()) {
-			refs = o.getObjRefPath();
-		} else {
-			refs = null;
+			count++;
 		}
-		if (refs != null && objrefs != null) {
-			throw new IllegalArgumentException("Only one of an object " +
-					"reference path or an object path may be specified");
+		if (o.getToObjPath() != null && !o.getToObjPath().isEmpty()) {
+			count++;
 		}
-		final List<ObjectIdentifier> ret = new LinkedList<ObjectIdentifier>();
-		if (refs != null) {
-			int refcount = 1;
-			for (final String r: refs) {
-				try {
-					ret.add(processObjectReference(r));
-				} catch (IllegalArgumentException e) {
-					throw new IllegalArgumentException(String.format(
-							"Invalid object reference (%s) at position #%s: %s",
-							r, refcount, e.getLocalizedMessage()), e);
-				}
-				refcount++;
+		if (o.getToObjRefPath() != null && !o.getToObjRefPath().isEmpty()) {
+			count++;
+		}
+		return count;
+	}
+
+	private static class RefChainResult {
+		
+		private boolean isToPath;
+		private List<ObjectIdentifier> path;
+		
+		public RefChainResult(
+				final List<ObjectIdentifier> path,
+				final boolean isToPath) {
+			this.isToPath = isToPath;
+			this.path = path;
+		}
+	}
+	
+	private static RefChainResult processRefChains(
+			final ObjectSpecification o) {
+		if (countRefPathsInObjectSpec(o) > 1) {
+			throw new IllegalArgumentException("Only one of the 5 options " +
+					"for specifying an object reference path is allowed");
+		}
+		if (o.getObjPath() != null && !o.getObjPath().isEmpty()) {
+			return new RefChainResult(processObjectPath(
+					o.getObjPath(), false), false);
+		} else if (o.getToObjPath() != null && !o.getToObjPath().isEmpty()) {
+			return new RefChainResult(processObjectPath(
+					o.getToObjPath(), true), true);
+		}
+		if (o.getObjRefPath() != null && !o.getObjRefPath().isEmpty()) {
+			return new RefChainResult(processRefPath(
+					o.getObjRefPath(), false), false);
+		} else if (o.getToObjRefPath() != null &&
+				!o.getToObjRefPath().isEmpty()) {
+			return new RefChainResult(processRefPath(
+					o.getToObjRefPath(), true), true);
+		}
+		return null;
+	}
+
+	private static List<ObjectIdentifier> processRefPath(
+			final List<String> objRefPath,
+			final boolean isToPath) {
+		int refcount = isToPath ? 1 : 2;
+		final List<ObjectIdentifier> ret = new LinkedList<>();
+		for (final String r: objRefPath) {
+			try {
+				ret.add(processObjectReference(r));
+			} catch (IllegalArgumentException e) {
+				throw new IllegalArgumentException(String.format(
+						"Invalid object reference (%s) at position #%s: %s",
+						r, refcount, e.getLocalizedMessage()), e);
 			}
+			refcount++;
 		}
-		if (objrefs != null) {
-			int refcount = 1;
-			for (final ObjectIdentity oi: objrefs) {
-				try {
-					ret.add(processObjectIdentifier(oi));
-				} catch (IllegalArgumentException e) {
-					throw new IllegalArgumentException(String.format(
-							"Invalid object id at position #%s: %s",
-							refcount, e.getLocalizedMessage()), e);
-				}
-				refcount++;
+		return ret;
+	}
+
+	private static List<ObjectIdentifier> processObjectPath(
+			final List<ObjectIdentity> objPath,
+			final boolean isToPath) {
+		int refcount = isToPath ? 1 : 2;
+		final List<ObjectIdentifier> ret = new LinkedList<>();
+		for (final ObjectIdentity oi: objPath) {
+			try {
+				ret.add(processObjectIdentifier(oi));
+			} catch (IllegalArgumentException e) {
+				throw new IllegalArgumentException(String.format(
+						"Invalid object id at position #%s: %s",
+						refcount, e.getLocalizedMessage()), e);
 			}
-		}
-		if (ret.isEmpty()) {
-			return null;
+			refcount++;
 		}
 		return ret;
 	}